+++ Changelog.md
@@ -13,6 +13,7 @@
 ### BugFix
 
 - Fix cross-bucket contamination in snapshot candidate query [#59](https://github.com/NEventStore/NEventStore.Persistence.SQL/issues/59)
+- Fix async paged-query infinite loop when PageSize is zero [#60](https://github.com/NEventStore/NEventStore.Persistence.SQL/issues/60)
 
 ### Breaking Changes